Time: 0 ms (100.00%), Space: 6 MB (24.89%) - LeetHub

+++ 69-sqrtx/69-sqrtx.cpp
@@ -6,7 +6,7 @@ public:
      int high=x;
         int ans;
         while(low<=high){
-           long long int mid=(low+high)/2;
+            long int mid=(low+high)/2;
             // cout<<mid<<endl;
             if(mid*mid==x) return mid;
             else if(mid*mid>x){
